+++ packages/app-preset/__tests__/package-metadata.test.js
@@ -4,5 +4,5 @@ const packageJson = require('../package.json');
 
 test('Oxy peers accept compatible patches without crossing a major', () => {
   assert.equal(packageJson.peerDependencies['@oxy.so/bloom'], '^2.0.0 || ^3.0.0 || ^4.2.0');
-  assert.equal(packageJson.peerDependencies['@oxy.so/services'], '^2.0.0 || ^3.0.0 || ^4.0.0');
+  assert.equal(packageJson.peerDependencies['@oxy.so/services'], '^2.0.0 || ^3.0.0 || ^4.0.0 || ^5.0.0');
 });